refactor!(cli): remove auto command
