+++ bumpwright/cli.py
@@ -354,116 +354,6 @@ def bump_command(args: argparse.Namespace) -> int:
     return 0
 
 
-def auto_command(args: argparse.Namespace) -> int:
-    """CLI command to decide and apply a version bump in one step.
-
-    When ``--base`` is omitted, the upstream of the current branch is used.
-    ``--head`` defaults to ``HEAD``.
-
-    Args:
-        args: Parsed command-line arguments.
-
-    Returns:
-        Process exit code.
-
-    Examples:
-        Show Markdown output and apply the bump:
-
-        $ bumpwright auto --format md
-        **bumpwright** suggests: `patch`
-
-        (no API-impacting changes detected)
-
-        Bumped version: 1.2.3 -> 1.2.4 (patch)
-
-        Emit JSON for automation:
-
-        $ bumpwright auto --format json
-        {
-          "level": "patch",
-          "impacts": [],
-          "old_version": "1.2.3",
-          "new_version": "1.2.4"
-        }
-
-    If no relevant files have changed or no API impacts are detected, the
-    command prints ``No version bump needed`` and exits without modifying the
-    project version.
-    """
-
-    base = args.base or last_release_commit() or _infer_base_ref()
-    head = args.head
-    cfg = load_config(args.config)
-
-    try:
-        pyproject = _resolve_pyproject(args.pyproject)
-    except FileNotFoundError as exc:
-        print(f"Error: {exc}", file=sys.stderr)
-        return 1
-
-    paths = args.version_path or cfg.version.paths
-    version_files = {p for p in paths if not any(ch in p for ch in "*?[")}
-
-    try:
-        changed = changed_paths(base, head)
-    except Exception:  # git diff may fail on initial commit
-        changed = None
-
-    if changed is not None:
-        filtered = {
-            p for p in changed if p != Path(pyproject).name and p not in version_files
-        }
-        if not filtered:
-            print("No version bump needed")
-            return 0
-
-    old_api = _build_api_at_ref(base, cfg.project.public_roots, cfg.ignore.paths)
-    new_api = _build_api_at_ref(head, cfg.project.public_roots, cfg.ignore.paths)
-    impacts = diff_public_api(
-        old_api, new_api, return_type_change=cfg.rules.return_type_change
-    )
-    impacts.extend(_run_analyzers(base, head, cfg))
-    level = decide_bump(impacts)
-    if level is None:
-        print("No version bump needed")
-        return 0
-
-    ignore = args.version_ignore or cfg.version.ignore
-    vc = apply_bump(
-        level,
-        pyproject_path=pyproject,
-        dry_run=args.dry_run,
-        paths=paths,
-        ignore=ignore,
-    )
-
-    if args.format == "json":
-        print(
-            json.dumps(
-                {
-                    "level": level,
-                    "impacts": [i.__dict__ for i in impacts],
-                    "old_version": vc.old,
-                    "new_version": vc.new,
-                },
-                indent=2,
-            )
-        )
-    elif args.format == "md":
-        print(f"**bumpwright** suggests: `{level}`\n")
-        print(_format_impacts_text(impacts))
-        print()
-        print(f"Bumped version: {vc.old} -> {vc.new} ({vc.level})")
-    else:
-        print(f"Suggested bump: {level}")
-        print(_format_impacts_text(impacts))
-        print(f"Bumped version: {vc.old} -> {vc.new} ({vc.level})")
-
-    if not args.dry_run:
-        _commit_tag(str(pyproject), vc.new, args.commit, args.tag)
-    return 0
-
-
 def main(argv: list[str] | None = None) -> int:
     """Entry point for the ``bumpwright`` CLI.
 
@@ -569,60 +459,6 @@ def main(argv: list[str] | None = None) -> int:
         help="Display the new version without modifying any files.",
     )
     p_bump.set_defaults(func=bump_command)
-
-    p_auto = sub.add_parser(
-        "auto",
-        help="Decide and apply bump, committing and tagging optionally",
-        description=(
-            "Combine 'decide' and 'bump' to infer the level and update files in one step."
-        ),
-    )
-    p_auto.add_argument(
-        "--base",
-        help="Base git reference; defaults to the upstream of the current branch.",
-    )
-    p_auto.add_argument(
-        "--head",
-        default="HEAD",
-        help="Head git reference; defaults to the current HEAD.",
-    )
-    p_auto.add_argument(
-        "--format",
-        choices=["text", "md", "json"],
-        default="text",
-        help="Output style: plain text, Markdown, or machine-readable JSON.",
-    )
-    p_auto.add_argument(
-        "--pyproject",
-        default="pyproject.toml",
-        help="Path to the project's pyproject.toml file.",
-    )
-    p_auto.add_argument(
-        "--version-path",
-        action="append",
-        dest="version_path",
-        help="Glob pattern for files that contain the project version (repeatable).",
-    )
-    p_auto.add_argument(
-        "--version-ignore",
-        action="append",
-        dest="version_ignore",
-        help="Glob pattern for paths to exclude from version updates.",
-    )
-    p_auto.add_argument(
-        "--commit",
-        action="store_true",
-        help="Create a git commit for the version change.",
-    )
-    p_auto.add_argument(
-        "--tag", action="store_true", help="Create a git tag for the new version."
-    )
-    p_auto.add_argument(
-        "--dry-run",
-        action="store_true",
-        help="Display the new version without modifying any files.",
-    )
-    p_auto.set_defaults(func=auto_command)
 
     args = parser.parse_args(argv)
     if not hasattr(args, "func"):
