+++ tests/test_cli_decide_default.py
@@ -4,15 +4,15 @@ import subprocess
 import sys
 from pathlib import Path
 
-from tests.test_cli_auto import _run, _setup_repo
+from tests.cli_helpers import run, setup_repo
 
 
 def test_decide_defaults_to_previous_commit(tmp_path: Path) -> None:
-    repo, pkg, _ = _setup_repo(tmp_path)
+    repo, pkg, _ = setup_repo(tmp_path)
 
     (pkg / "extra.py").write_text("def bar() -> int:\n    return 2\n", encoding="utf-8")
-    _run(["git", "add", "pkg/extra.py"], repo)
-    _run(["git", "commit", "-m", "feat: add bar"], repo)
+    run(["git", "add", "pkg/extra.py"], repo)
+    run(["git", "commit", "-m", "feat: add bar"], repo)
 
     res = subprocess.run(
         [
